requirements.txt
pdfminer.six
openai
json
os
